Factor out scala.meta.internal.scalacp
